recursive-xslt.xpl updated
Modified to allow some sensible functionality with the use of secondary outputs on xslt steps. Primary output and secondary and  now converted to a sequence and the sequence is processed via any further stylesheets.

This may be revisited if I can think of a better way to handle secondary outputs that lead to secondary outputs and so on as their seems to be no clear logical approach to keeping the results separate as we may get logarithmic growth in the number of outputs (in theory)

Tests updated.

Signed-off-by: Nic Gibson <nicg@corbas.net>